requirements.txt
addict==2.4.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
coacd==1.0.7
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.2
cupoch==0.2.11.0
cycler==0.12.1
dash==3.2.0
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
Flask==3.1.2
fonttools==4.59.2
fsspec==2025.9.0
idna==3.10
importlib_metadata==8.7.0
ipython==8.37.0
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lxml==6.0.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy_extensions==1.1.0
narwhals==2.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
open3d==0.19.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11==3.0.1
pyglet==1.5.31
pygltflib==1.16.5
Pygments==2.19.2
pyparsing==3.2.3
pyperclip==1.9.0
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2025.2
pyVHACD==0.0.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
retrying==1.4.2
rpds-py==0.27.1
sapien==3.0.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
traitlets==5.14.3
transforms3d==0.4.2
trimesh==4.8.1
triton==3.4.0
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.3
zipp==3.23.0
